extract state
